Moved the ServiceMetrics constructor and definition into its own file. Updated index to use new file
